global interrupts HANDLED

+++ src/MCAL/DMA/dma.c
@@ -15,6 +15,7 @@
 #include "platform_types.h"
 #include "rcc.h"
 #include "mcal_def.h"
+#include "nvic.h"
 
 
 /*------------------- DMA static enums ----------------------*/
@@ -461,52 +462,52 @@ void DMA_Start_IT(DMA_HandleTypeDef *streamManager, uint32_t SrcAddress, uint32_
 
 		/* Enable DMA global interrupts */
 		switch((uint32)streamManager->Instance){
-			case DMA1_Stream0:
+			case (uint32)DMA1_Stream0:
 				__NVIC_EnableIRQ(DMA1_Stream0_IRQn);
 				break;
-			case DMA1_Stream1:
+			case (uint32)DMA1_Stream1:
 				__NVIC_EnableIRQ(DMA1_Stream1_IRQn);
 				break;
-			case DMA1_Stream2:
+			case (uint32)DMA1_Stream2:
 				__NVIC_EnableIRQ(DMA1_Stream2_IRQn);
 				break;
-			case DMA1_Stream3:
+			case (uint32)DMA1_Stream3:
 				__NVIC_EnableIRQ(DMA1_Stream3_IRQn);
 				break;
-			case DMA1_Stream4:
+			case (uint32)DMA1_Stream4:
 				__NVIC_EnableIRQ(DMA1_Stream4_IRQn);
 				break;
-			case DMA1_Stream5:
+			case (uint32)DMA1_Stream5:
 				__NVIC_EnableIRQ(DMA1_Stream5_IRQn);
 				break;
-			case DMA1_Stream6:
+			case (uint32)DMA1_Stream6:
 				__NVIC_EnableIRQ(DMA1_Stream6_IRQn);
 				break;
-			case DMA1_Stream7:
+			case (uint32)DMA1_Stream7:
 				__NVIC_EnableIRQ(DMA1_Stream7_IRQn);
 				break;
-			case DMA2_Stream0:
+			case (uint32)DMA2_Stream0:
 				__NVIC_EnableIRQ(DMA2_Stream0_IRQn);
 				break;
-			case DMA2_Stream1:
+			case (uint32)DMA2_Stream1:
 				__NVIC_EnableIRQ(DMA2_Stream1_IRQn);
 				break;
-			case DMA2_Stream2:
+			case (uint32)DMA2_Stream2:
 				__NVIC_EnableIRQ(DMA2_Stream2_IRQn);
 				break;
-			case DMA2_Stream3:
+			case (uint32)DMA2_Stream3:
 				__NVIC_EnableIRQ(DMA2_Stream3_IRQn);
 				break;
-			case DMA2_Stream4:
+			case (uint32)DMA2_Stream4:
 				__NVIC_EnableIRQ(DMA2_Stream4_IRQn);
 				break;
-			case DMA2_Stream5:
+			case (uint32)DMA2_Stream5:
 				__NVIC_EnableIRQ(DMA2_Stream5_IRQn);
 				break;
-			case DMA2_Stream6:
+			case (uint32)DMA2_Stream6:
 				__NVIC_EnableIRQ(DMA2_Stream6_IRQn);
 				break;
-			case DMA2_Stream7:
+			case (uint32)DMA2_Stream7:
 				__NVIC_EnableIRQ(DMA2_Stream7_IRQn);
 				break;
